Create getCurrentPlayer
Looking over the code, I decided to remove the currentPlayer variable in Gameboard and instead create a new boolean property for each CreatePlayer object.

I made this decision because it felt more intuitive and cleaner not to use changeCurrentPlayer to get the currentPlayer argument for createSymbol. To return the currentPlayer for createSymbol I created a function that does only that, return the player with the property isCurrentPlayer set to true.